+++ lib/src/extensions/extensions.dart
@@ -1,8 +1,9 @@
+export 'color_extension.dart';
 export 'context_extension.dart';
 export 'datetime_extension.dart';
 export 'edge_inset_extension.dart';
 export 'file_extension.dart';
 export 'iterable_extension.dart';
 export 'num_extension.dart';
-export 'widget_extension.dart';
 export 'string_extension.dart';
+export 'widget_extension.dart';